requirements.txt
novyx>=0.1.5
python-dotenv
